+++ src/lib.rs
@@ -114,6 +114,7 @@ mod frequency;
 mod limit;
 pub mod scaling;
 mod spectrum;
+mod util;
 pub mod windows;
 
 // test module for large "integration"-like tests